Move `ReservationQuerySet` to its own file
